refactor: reuse departure time
