+++ src/modules/trip-details/index.ts
@@ -1,4 +1,3 @@
 export { default as TripRow } from './trip-row';
 export { default as DecorationLine } from './decoration-line';
 export { default as useRealtimeText } from './use-realtime-text';
-export { default as DepartureTime } from './departure-time';
