+++ src/page-modules/departures/stop-place/stop-place.module.css
@@ -97,9 +97,6 @@
   top: calc(var(--spacings-small) * -1);
   left: calc(var(--spacings-small) * -1);
 }
-.departureTime {
-  text-align: right;
-}
 
 @media only screen and (max-width: 650px) {
   .stopPlaceContainer {
